requirements.txt
beautifulsoup4==4.12.2
fastapi==0.95.2
faiss-cpu==1.7.4
html2text==2020.1.16
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
Jinja2==3.1.2
tiktoken==0.3.3
langchain==0.0.254
llama-index==0.6.10.post1
openai==0.27.5
python-dotenv==0.16.0
selenium==4.9.1
uvicorn==0.22.0
webdriver-manager==3.8.6